Moving some code around. Renaming files to be more reflective of what they do currently. Refactoring comingled code into separate files.
