+++ flang/lib/Optimizer/CMakeLists.txt
@@ -1,7 +1,3 @@
-if ((CMAKE_CXX_COMPILER_ID MATCHES "Clang"))
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-unused-private-field")
-endif()
-
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-unused-parameter -Wno-error")
 
 add_subdirectory(Analysis)
