+++ flang/lib/Optimizer/Transforms/CMakeLists.txt
@@ -1,10 +1,11 @@
 get_property(dialect_libs GLOBAL PROPERTY MLIR_DIALECT_LIBS)
 
 add_flang_library(FIRTransforms
+  ControlFlowConverter.cpp
   CSE.cpp
   MemToReg.cpp
+  RaiseToAffine.cpp
   RewriteLoop.cpp
-  StdConverter.cpp
 
   DEPENDS
   FIROpsIncGen
