+++ flang/tools/tco/tco.cpp
@@ -95,7 +95,7 @@ static int compileFIR() {
     pm.addPass(fir::createPromoteToAffinePass());
     // convert fir dialect to loop
     pm.addPass(fir::createLowerToLoopPass());
-    pm.addPass(fir::createFIRToStdPass(kindMap));
+    pm.addPass(fir::createControlFlowLoweringPass());
     // convert loop dialect to standard
     pm.addPass(mlir::createLowerToCFGPass());
     //pm.addPass(fir::createMemToRegPass());
